pyproject.toml
[tool.poetry]
name = "langchain-rag"
version = "0.1.0"
description = "RAG built using Langchain for 자란다 국책과제"
authors = ["domini04 <sinyubi04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
streamlit = "^1.37.1"
langchain-core = "^0.2.29"
langchain-community = "^0.2.11"
langgraph = "^0.2.3"
langserve = "^0.2.2"
langsmith = "^0.1.99"
torch = "^2.4.0"
accelerate = "^0.33.0"
transformers = "^4.44.0"
notebook = "^7.2.1"
python-dotenv = "^1.0.1"
blinker = "^1.8.2"
langchain-openai = "^0.1.21"
langchain-unstructured = "^0.1.1"
langchain-text-splitters = "^0.2.2"
sentence-transformers = "^3.0.1"
langchain-huggingface = "^0.0.3"
faiss-cpu = "^1.8.0.post1"
rank-bm25 = "^0.2.2"
pypdf = "^4.3.1"
langchain-cli = "^0.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 